# aes-soft: use fixslicing for AES encryption