… aes/aes-soft/src/fixslice.rs
… aes/aes-soft/tests/lib.rs