@tarcieri 2020-10-26
Now with AES-192 (thanks @peterdettman!)

```
test aes192_decrypt  ... bench:         713 ns/iter (+/- 94) = 22 MB/s
test aes192_decrypt8 ... bench:       3,651 ns/iter (+/- 350) = 35 MB/s
test aes192_encrypt  ... bench:         338 ns/iter (+/- 73) = 47 MB/s
test aes192_encrypt8 ... bench:       1,355 ns/iter (+/- 79) = 94 MB/s
```