@tarcieri 2020-10-26
With AES-192 supported, I was able to continue using the macros to define the `Aes128`, `Aes192`, and `Aes256` types, which makes me feel a lot better about this PR (i.e. fixslicing is used across the board for encryption, and bitslicing is used for decryption).